[FEAT] Add Stage Indicator, TimeOver Logic
스테이지 표시 텍스트 추가 및
타임 오버시 스테이지 리셋 기능 추가

+++ Assets/Scripts/UIs/GoldIndicator.cs
@@ -15,6 +15,6 @@ public class GoldIndicator : MonoBehaviour
 
     private void UpdateGoldUI()
     {
-        goldText.text = GameManager.Instance.player.gold.ToString();
+        goldText.text = ((int)GameManager.Instance.player.gold).ToString();
     }
 }
